fix(bitcoin): remove .int() from all amount_btc Zod schemas
Five schemas used z.number().int() on amount_btc fields, rejecting every
fractional BTC value (e.g. 0.001 BTC fails .int()). These are sats-era
leftovers from before the sats→BTC migration.

Changes:
- transactions/route: remove .int(), cap at 21M BTC (supply cap)
- proposals validation: remove .int().min(1) — required 1+ whole BTC
- ai-credits/add: remove .int().min(1).max(1000000 sats)
- project support validation: remove .int()
- withdrawal route: remove .int(), replace sats-based min with MIN_WITHDRAWAL_BTC
- WithdrawDialog: fix UI sending sats as amount_btc — divide by 100M
- types.ts: add MIN_WITHDRAWAL_BTC constant (1000 sats in BTC)
- events/route: clean up lint warning from previous destructure pattern

+++ src/services/projects/support/validation.ts
@@ -22,7 +22,7 @@ export const supportProjectSchema = z
     support_type: supportTypeSchema,
 
     // Bitcoin donation fields
-    amount_btc: z.number().positive().int().optional(),
+    amount_btc: z.number().positive().optional(),
     lightning_invoice: z.string().optional(),
     transaction_hash: z.string().optional(),
